// src/api/models/webConfig.ts
/**
 * Generated by orval v8.0.0-rc.0 🍺
 * Do not edit manually.
 * VCCI
 * Coded by Meu TEAM
 * OpenAPI spec version: 1.0.0
 */

export interface WebConfig {
  id?: string;
  name?: string;
  address?: string;
  name_en?: string;
  address_en?: string;
  logo?: string;
  link?: string;
  phone?: string;
  created_at?: string;
  updated_at?: string;
  created_by?: string;
  updated_by?: string;
  email?: string;
  social?: string;
}